+++ src/main/java/org/truffleruby/core/format/printf/PrintfSimpleTreeBuilder.java
@@ -14,7 +14,7 @@ import org.jcodings.specific.USASCIIEncoding;
 import org.truffleruby.RubyContext;
 import org.truffleruby.core.format.FormatNode;
 import org.truffleruby.core.format.LiteralFormatNode;
-import org.truffleruby.core.format.control.SequenceNode;
+import org.truffleruby.core.format.SharedTreeBuilder;
 import org.truffleruby.core.format.convert.ToDoubleWithCoercionNodeGen;
 import org.truffleruby.core.format.convert.ToIntegerNodeGen;
 import org.truffleruby.core.format.convert.ToStringNodeGen;
@@ -193,7 +193,7 @@ public class PrintfSimpleTreeBuilder {
 
     public FormatNode getNode() {
         buildTree();
-        return new SequenceNode(sequence.toArray(new FormatNode[sequence.size()]));
+        return SharedTreeBuilder.createSequence(sequence.toArray(new FormatNode[sequence.size()]));
     }
 
 }