+++ src/main/java/org/truffleruby/core/format/read/bytes/ReadBinaryStringNode.java
@@ -20,6 +20,7 @@ import org.truffleruby.core.format.FormatNode;
 import org.truffleruby.core.format.read.SourceNode;
 import org.truffleruby.core.rope.CodeRange;
 import org.truffleruby.core.string.StringNodes;
+import org.truffleruby.language.objects.TaintNode;
 
 import java.util.Arrays;
 
@@ -56,7 +57,8 @@ public abstract class ReadBinaryStringNode extends FormatNode {
 
     @Specialization
     public DynamicObject read(VirtualFrame frame, byte[] source,
-                              @Cached("create()") StringNodes.MakeStringNode makeStringNode) {
+                              @Cached("create()") StringNodes.MakeStringNode makeStringNode,
+                              @Cached("create()") TaintNode taintNode) {
         final int start = getSourcePosition(frame);
 
         int length;
@@ -109,7 +111,13 @@ public abstract class ReadBinaryStringNode extends FormatNode {
 
         setSourcePosition(frame, start + length);
 
-        return makeStringNode.executeMake(Arrays.copyOfRange(source, start, start + usedLength), ASCIIEncoding.INSTANCE, CodeRange.CR_UNKNOWN);
+        final DynamicObject string = makeStringNode.executeMake(Arrays.copyOfRange(source, start, start + usedLength), ASCIIEncoding.INSTANCE, CodeRange.CR_UNKNOWN);
+
+        if (isSourceTainted(frame)) {
+            taintNode.executeTaint(string);
+        }
+
+        return string;
     }
 
     private int indexOfFirstNull(byte[] bytes, int start, int length) {