+++ src/main/java/org/truffleruby/core/string/StringNodes.java
@@ -1,4 +1,3 @@
-
 /*
  * Copyright (c) 2013, 2018 Oracle and/or its affiliates. All rights reserved. This
  * code is released under a tri EPL/GPL/LGPL license. You can use it,
@@ -147,6 +146,8 @@ import org.truffleruby.language.control.RaiseException;
 import org.truffleruby.language.dispatch.CallDispatchHeadNode;
 import org.truffleruby.language.objects.AllocateObjectNode;
 import org.truffleruby.language.objects.IsTaintedNode;
+import org.truffleruby.language.objects.ReadObjectFieldNode;
+import org.truffleruby.language.objects.ReadObjectFieldNodeGen;
 import org.truffleruby.language.objects.TaintNode;
 import org.truffleruby.language.yield.YieldNode;
 
@@ -2443,7 +2444,7 @@ public abstract class StringNodes {
 
     @NodeChild(value = "string", type = RubyNode.class)
     @NodeChild(value = "format", type = RubyNode.class)
-    @CoreMethod(names = "unpack", required = 1, taintFrom = 1)
+    @CoreMethod(names = "unpack", required = 1)
     @ImportStatic({ StringCachingGuards.class, StringOperations.class })
     public abstract static class UnpackNode extends CoreMethodNode {
 
@@ -2465,14 +2466,16 @@ public abstract class StringNodes {
                 @Cached("privatizeRope(format)") Rope cachedFormat,
                 @Cached("create(compileFormat(format))") DirectCallNode callUnpackNode,
                 @Cached("create()") RopeNodes.BytesNode bytesNode,
-                @Cached("create()") RopeNodes.EqualNode equalNode) {
+                @Cached("create()") RopeNodes.EqualNode equalNode,
+                @Cached("create()") IsTaintedNode isTaintedNode,
+                @Cached("createReadAssociatedNode()") ReadObjectFieldNode readAssociatedNode) {
             final Rope rope = rope(string);
 
             final ArrayResult result;
 
             try {
                 result = (ArrayResult) callUnpackNode.call(
-                        new Object[]{ bytesNode.execute(rope), rope.byteLength() });
+                        new Object[]{ bytesNode.execute(rope), rope.byteLength(), isTaintedNode.executeIsTainted(string), readAssociatedNode.execute(string) });
             } catch (FormatException e) {
                 exceptionProfile.enter();
                 throw FormatExceptionTranslator.translate(this, e);
@@ -2486,20 +2489,26 @@ public abstract class StringNodes {
                 DynamicObject string,
                 DynamicObject format,
                 @Cached("create()") IndirectCallNode callUnpackNode,
-                @Cached("create()") RopeNodes.BytesNode bytesNode) {
+                @Cached("create()") RopeNodes.BytesNode bytesNode,
+                @Cached("create()") IsTaintedNode isTaintedNode,
+                @Cached("createReadAssociatedNode()") ReadObjectFieldNode readAssociatedNode) {
             final Rope rope = rope(string);
 
             final ArrayResult result;
 
             try {
                 result = (ArrayResult) callUnpackNode.call(compileFormat(format),
-                        new Object[]{ bytesNode.execute(rope), rope.byteLength() });
+                        new Object[]{ bytesNode.execute(rope), rope.byteLength(), isTaintedNode.executeIsTainted(string), readAssociatedNode.execute(string) });
             } catch (FormatException e) {
                 exceptionProfile.enter();
                 throw FormatExceptionTranslator.translate(this, e);
             }
 
             return finishUnpack(result);
+        }
+
+        protected ReadObjectFieldNode createReadAssociatedNode() {
+            return ReadObjectFieldNodeGen.create(Layouts.ASSOCIATED_IDENTIFIER, null);
         }
 
         private DynamicObject finishUnpack(ArrayResult result) {